Create OpaqueEnvelope class and tests
Summary:
OpaqueEnvelope is a wrapper class to pass around sensitive strings, such as passwords, encryption keys, etc. The "key" is generated once per request and never saved, so the internal masked value is not meaningful across requests (any sort of serialization effort is meaningless. Further, the `__toString` and `__debugInfo` magic methods help prevent the maksed value from ever showing up.

This is not intended to be crytographically secure, nor as a replacement for storing keys and passwords safely - it is simply another tool when following Defense In Depth principles.

Inspired by and derived from the Arcanist library of the same name, with some >=PHP5.3 additions.

Test Plan: Unit tests included, which pass

Differential Revision: https://phabricator.ericstern.com/D50